requirements.txt
Django~=3.1.5
djangorestframework~=3.12.2
asgiref==3.3.4
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
